re-organize clinical variable
