+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/EnhancedContinuous.js
@@ -1,0 +1,162 @@
+import {
+  compose,
+  setDisplayName,
+  withProps,
+  withPropsOnChange,
+  withState,
+} from 'recompose';
+import { connect } from 'react-redux';
+import { isEqual } from 'lodash';
+
+import { withTheme } from '@ncigdc/theme';
+import {
+  // humanify,
+  createFacetFieldString,
+} from '@ncigdc/utils/string';
+import { setModal } from '@ncigdc/dux/modal';
+
+import ContinuousCustomBinsModal from './modals/ContinuousCustomBinsModal';
+import RecomposeUtils from './helpers';
+import EnhancedShared from './EnhancedShared';
+
+export default compose(
+  setDisplayName('EnhancedContinuousVariableCard'),
+  connect((state: any) => ({ analysis: state.analysis })),
+  withTheme,
+  RecomposeUtils,
+  withState('qqData', 'setQQData', []),
+  withState('qqDataIsSet', 'setQQDataIsSet', false),
+  withProps(({ data: { explore }, fieldName }) => {
+    const dataStats = explore
+      ? explore.cases.aggregations[
+        `${createFacetFieldString(fieldName)}`].stats
+      : {
+        Max: null,
+        Min: null,
+      };
+
+    const defaultMin = dataStats.Min;
+    const defaultMax = dataStats.Max + 1;
+    // api excludes the max number
+
+    const defaultQuarter = (defaultMax - defaultMin) / 4;
+
+    const defaultNumberOfBins = 5;
+    const defaultBinSize = (defaultMax - defaultMin) /
+      defaultNumberOfBins;
+
+    const defaultBins = Array(defaultNumberOfBins).fill(1)
+      .map((val, key) => {
+        const from = key * defaultBinSize + defaultMin;
+        const to = (key + 1) === defaultNumberOfBins
+          ? defaultMax
+          : (defaultMin + (key + 1) * defaultBinSize);
+        const objKey = `${from}-${to}`;
+
+        return ({
+          [objKey]: {
+            key: objKey,
+          },
+        });
+      }).reduce((acc, curr) => Object.assign({}, acc, curr), {});
+
+    return ({
+      defaultData: {
+        bins: defaultBins,
+        max: defaultMax,
+        min: defaultMin,
+        quarter: defaultQuarter,
+      },
+    });
+  }),
+  withPropsOnChange(
+    (props, nextProps) =>
+      props.variable.continuousBinType !== nextProps.variable.continuousBinType,
+    ({ variable: { continuousBinType } }) => ({
+      resetBinsDisabled: continuousBinType === 'default',
+    })
+  ),
+  withPropsOnChange(
+    (props, nextProps) => !isEqual(
+      props.defaultData,
+      nextProps.defaultData
+    ) ||
+      !isEqual(
+        props.variable.customInterval,
+        nextProps.variable.customInterval
+      ) ||
+      !isEqual(
+        props.variable.customRanges,
+        nextProps.variable.customRanges
+      ),
+    ({
+      defaultData,
+      dispatch,
+      dispatchUpdateClinicalVariable,
+      fieldName,
+      variable,
+    }) => ({
+      openCustomBinModal: () => dispatch(setModal(
+        <ContinuousCustomBinsModal
+          continuousBinType={variable.continuousBinType}
+          customInterval={variable.customInterval}
+          customRanges={variable.customRanges}
+          defaultData={defaultData}
+          fieldName={humanify({ term: fieldName })}
+          onClose={() => dispatch(setModal(null))}
+          onUpdate={(
+            newBins,
+            continuousBinType,
+            customInterval,
+            customRanges,
+            continuousReset,
+          ) => {
+            dispatchUpdateClinicalVariable({
+              value: continuousReset
+                ? defaultData.bins
+                : newBins,
+              variableKey: 'bins',
+            });
+            dispatchUpdateClinicalVariable({
+              value: continuousReset
+                ? 'default'
+                : continuousBinType,
+              variableKey: 'continuousBinType',
+            });
+            !continuousReset &&
+              continuousBinType === 'interval' &&
+              (
+                dispatchUpdateClinicalVariable({
+                  value: customInterval,
+                  variableKey: 'customInterval',
+                })
+              );
+            !continuousReset &&
+              continuousBinType === 'range' &&
+              (
+                dispatchUpdateClinicalVariable({
+                  value: customRanges,
+                  variableKey: 'customRanges',
+                })
+              );
+            continuousReset &&
+              (
+                dispatchUpdateClinicalVariable({
+                  value: [],
+                  variableKey: 'customRanges',
+                })
+              );
+            continuousReset &&
+              (
+                dispatchUpdateClinicalVariable({
+                  value: {},
+                  variableKey: 'customInterval',
+                })
+              );
+            dispatch(setModal(null));
+          }}
+          />
+      )),
+    })
+  ),
+)(EnhancedShared);
