+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/index.js
@@ -1,4 +1,4 @@
-import CategoricalVariableCard from './CategoricalVariableCard';
-import ContinuousVariableCard from './ContinuousCard';
+import CategoricalVariableCard from './EnhancedCategorical';
+import ContinuousVariableCard from './EnhancedContinuous';
 
 export { CategoricalVariableCard, ContinuousVariableCard };
