+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/modals/ContinuousCustomBinsModal/index.js
@@ -1,0 +1,3 @@
+import ContinuousCustomBinsModal from './ContinuousCustomBinsModal';
+
+export default ContinuousCustomBinsModal;
